Changes made to allow code to compile with a modern windows compiler and still be compatible with Linux/Unix/OSX. Will check in a new Visual C++ 2008 directory in a separte commit.
made frequent use of,
#ifdef _MSC_VER
  <do microsoft specific stuff>
#else
  <do Linux/Unix/OSX stuff>
#endif

used static_cast<...>() to convert integer or longs to doubles in pow() and sqrt().

Added an assert(0) to the windows section of code that should never be called as it will lead to infinite recursion.

In a write() functionk, conveted some floats to doubles, but need to verify that they read in correctly, or change read code to double.

Changed a time variable from long to time_t.