Moved 6 functions out of HSprop.cc and into their own class, Evolve.  The Evolve classs is a containter for methods but has no data.
I tried using a namespace but ran into trouble on OSX.
This is basically the same thing.  Calls to the specific
version of evolve, prop, prop_ip, evolve_ip need to be
prefixed with an  Evolve:: specification.

e.g. Evolve::evolve(a, b, c);

In this particular case, I believe this solution is best as
it tells the user they are NOT using a local version of this
function, but one from the "Evolve::" class.  This name is
so overloaded that the clarity, in this case, helps (well,
at least I think so (DCT)).

Propogated this change by adding new include statments:
#include <HSLib/Evolve.h>;
and by changing "evolve()" to "Evolve::evolve()", etc, 
where appropriate, in the 12 or so classes.


Also, added declarations of non-member functions into 
header files, as needed. Previously they were only declared
in friend declarations that are in the class, but these are
no longer adequate as function declarations.