Added new dbwrite routine from lona as "dbwrite", and change previous dbwrite to dbwrite_old.
Modified both dbwrite and dbwrite_old to take an additional
variable (a vector of strings) that represents the file's header 
strings.    

Modified .i file to support using this vector of strings in python.  

The easiest way to do this is by creating a tuple of strings
or a list of strings and passing them to this function. 

The swig code handles the rest.