Fixed bugs in the OSX Makefile that caused `make test` to fail.
This fixes gamma ticket 7.

The main problem is that the lines that built the test apps wanted to use
static linking, and OS X is not keen on static linking. Removing -static 
from the target $(BINDIR)/runtests and friends allows all 19 tests to complete
with 0 failures.

Also, there was a hardcoded reference to primal/Linux in the OS X Makefile.
Removing this reference didn't break anything, so that's what I did.