Quick update to 4.2.1 which contained bad prebuilt libs for Python 2.6 and 2.7 under OS X.
gcc under OS X 10.6 builds 64 bit objects by default which is what it did
when I built PyGAMMA yesterday. I mistakenly believed that the build would 
fail loudly if I built 64-bit objects by mistake. I thought linking against
32-bit Python would cause swig's link step to fail. But under OS X we build
a dynamically linked _pygamma.so (the strongly encouraged default under OS X)
so linking doesn't happen until runtime. The build succeeds but produces a
64-bit library that won't run against 32-bit Python.

Ironically, the Python 2.5 _pygamma.so was spared this fate because it 
won't build at all in 64-bit mode and I had to hardcode -m32 in the Makefile
to get it to build. (This was a clue that I didn't pick up on.) Details:
http://scion.duhs.duke.edu/vespa/gamma/ticket/13

In order to ensure the correct -arch flag is passed to gcc, I added code
to the Makefile to get python's architecture. This fixes the broken libs
and ticket #13.

Obviously, I rebuilt the Python 2.6 and 2.7 libs for OS X. The 2.5 lib was
already OK.