Swigged MagVec, Bloch and BlochSys
Needed to redefine the  += and -=
operators in MagVec.cc.

There was a function missing from Bloch.cc, i.e.
matrix Mss_vector(...) 
and it was eliminated from the .h and .i files.

Another function, with a similar signature,
matrix Mo_vector(int argc, char* argv[], matrix& Meq, int& qn)
was not declared but was defined.
It was added to the .h and .i files.