Changed GenOp so that all the functions of type /=, +=, etc. return a "gen_op &". I had already done this for some of those functions, but not all of them.
Changed complex to add two new function:
c.real()
c.imag()
to return a 'double' and not 'double &' or 'double *'.
This will work better in Python as python treats
doubles as value types and not like objects and so
users actually get are copied double back and
not a pointer to a double.

PulGauss.i is only here because of some line termination
changes from windows to Linux format.

These changes were required to get gprofile3.py to work.