Small bug fixes.
- As reported by Rupert Kolb, the Linux Makefile was creating a bad symlink in 
the build directory for libgamma.so. The same syntax for 'ln -s' works under
OS X, so there's apparently differences in syntax for the ln command on these
systems. Anyway, the problem is fixed.

- The Linux build directory was called i686-pc-linux-gnu, implying it was 32-bit
regardless of whether it actually was or not. It's now just called linux-gnu.

- Fixed a bug where the OS X & Linux Makefiles were chmod-ing the wrong directory
after creating the PYSGPREBUILT directory. I don't think the chmod has any effect 
anyway so this was likely a harmless bug. I didn't make the change in the Solaris
Makefiles since there's no equivalent code there. We've never tried building 
PyGAMMA there as far as I know, and we don't have a way to test it.